Incomplete version: * Added Game, Objects, Collision, Map and Characters files * Replaced certain functions into those files
