+++ Characters.py
@@ -1,0 +1,5 @@
+import Objects
+
+
+class Character(Object):
+    etc = 0  # Todo
