+++ Collision.py
@@ -1,0 +1,11 @@
+class CollisionParams:
+    shape = 0  # Todo: Add sphere collision etc?
+    x = 0  # X offset of collision box in tiles (relative to parent object)
+    y = 0  # Y offset of collision box in tiles
+    width = 0  # Width of collision box in tiles
+    height = 0  # Height of collision box in tiles
+    solid = False
+
+    def __init__(self):
+        self.width = 0
+        self.height = 0
