+++ Map.py
@@ -1,0 +1,3 @@
+# Temporary file to identify tile size
+class MapClass:
+    TILE_SIZE = 80  # size of game tiles in pixels
